feat(web): move mission creation and schedule editing to full pages
The creation dialog crammed goal, kind, scheduling, and advanced
routing into one modal column; editing a schedule lived in a second
near-duplicate dialog. A shared MissionForm on /missions/new and
/missions/schedules/{id}/edit gives the form room: kind as described
cards, run-once/repeat as a segmented control, schedule fields in a
grid, and one submission path for create and edit.
